requirements.txt
torch>=2.9.0
torchvision>=0.20.0
streamlit>=1.28.1
opencv-python>=4.8.1.78
numpy>=1.24.3
pillow>=10.0.1
matplotlib>=3.7.2
scikit-learn>=1.3.0
google-generativeai>=0.3.0
